Tidy up testing
